+++ cluster/cluster.go
@@ -2,7 +2,7 @@ package cluster
 
 import (
 	"context"
-	"io/ioutil"
+	"fmt"
 	"math/rand"
 	"net"
 	"sort"
@@ -32,15 +32,57 @@ type Peer struct {
 	stopc  chan struct{}
 	readyc chan struct{}
 
+	peerLock    sync.RWMutex
+	peers       map[string]peer
+	failedPeers []peer
+
 	logger log.Logger
 }
 
+// peer is an internal type used for bookkeeping. It holds the state of peers
+// in the cluster.
+type peer struct {
+	status    PeerStatus
+	leaveTime time.Time
+
+	*memberlist.Node
+}
+
+// PeerStatus is the state that a peer is in.
+type PeerStatus int
+
 const (
-	DefaultPushPullInterval = 60 * time.Second
-	DefaultGossipInterval   = 200 * time.Millisecond
-	DefaultTcpTimeout       = 10 * time.Second
-	DefaultProbeTimeout     = 500 * time.Millisecond
-	DefaultProbeInterval    = 1 * time.Second
+	StatusNone PeerStatus = iota
+	StatusAlive
+	StatusLeaving
+	StatusLeft
+	StatusFailed
+)
+
+func (s PeerStatus) String() string {
+	switch s {
+	case StatusNone:
+		return "none"
+	case StatusAlive:
+		return "alive"
+	case StatusLeaving:
+		return "leaving"
+	case StatusLeft:
+		return "left"
+	case StatusFailed:
+		return "failed"
+	default:
+		panic(fmt.Sprintf("unknown PeerStatus: %d", s))
+	}
+}
+
+const (
+	DefaultPushPullInterval  = 60 * time.Second
+	DefaultGossipInterval    = 200 * time.Millisecond
+	DefaultTcpTimeout        = 10 * time.Second
+	DefaultProbeTimeout      = 500 * time.Millisecond
+	DefaultProbeInterval     = 1 * time.Second
+	DefaultReconnectInterval = 10 * time.Second
 )
 
 func Join(
@@ -55,6 +97,7 @@ func Join(
 	tcpTimeout time.Duration,
 	probeTimeout time.Duration,
 	probeInterval time.Duration,
+	reconnectInterval time.Duration,
 ) (*Peer, error) {
 	bindHost, bindPortStr, err := net.SplitHostPort(bindAddr)
 	if err != nil {
@@ -106,6 +149,7 @@ func Join(
 		stopc:  make(chan struct{}),
 		readyc: make(chan struct{}),
 		logger: l,
+		peers:  map[string]peer{},
 	}
 	p.delegate = newDelegate(l, reg, p)
 
@@ -120,7 +164,7 @@ func Join(
 	cfg.TCPTimeout = tcpTimeout
 	cfg.ProbeTimeout = probeTimeout
 	cfg.ProbeInterval = probeInterval
-	cfg.LogOutput = ioutil.Discard
+	cfg.LogOutput = &logWriter{l: l}
 
 	if advertiseAddr != "" {
 		cfg.AdvertiseAddr = advertiseHost
@@ -143,7 +187,121 @@ func Join(
 	if n > 0 {
 		go p.warnIfAlone(l, 10*time.Second)
 	}
+
+	if reconnectInterval != 0 {
+		go p.handleReconnect(reconnectInterval)
+	}
+	// TODO: Add a gc interval for failedPeers.
+
 	return p, nil
+}
+
+type logWriter struct {
+	l log.Logger
+}
+
+func (l *logWriter) Write(b []byte) (int, error) {
+	return len(b), level.Debug(l.l).Log("memberlist", string(b))
+}
+
+func (p *Peer) handleReconnect(d time.Duration) {
+	tick := time.NewTicker(d)
+	defer tick.Stop()
+
+	for {
+		select {
+		case <-p.stopc:
+			return
+		case <-tick.C:
+			p.reconnect()
+		}
+	}
+}
+
+func (p *Peer) reconnect() {
+	p.peerLock.RLock()
+	defer p.peerLock.RUnlock()
+
+	logger := log.With(p.logger, "msg", "reconnect")
+	for _, pr := range p.failedPeers {
+		// No need to do book keeping on failedPeers here. If a
+		// reconnect is successful, they will be announced in
+		// peerJoin().
+		if _, err := p.mlist.Join([]string{pr.Address()}); err != nil {
+			// TODO: failedReconnectionsCounter
+			level.Debug(logger).Log("result", "failure", "peer", pr.Node, "addr", pr.Address())
+		} else {
+			// TODO: successfulReconnectionsCounter
+			level.Debug(logger).Log("result", "success", "peer", pr.Node, "addr", pr.Address())
+		}
+	}
+}
+
+func (p *Peer) peerJoin(n *memberlist.Node) {
+	p.peerLock.RLock()
+	defer p.peerLock.RUnlock()
+
+	var oldStatus PeerStatus
+	pr, ok := p.peers[n.Name]
+	if !ok {
+		oldStatus = StatusNone
+		pr = peer{
+			status: StatusAlive,
+			Node:   n,
+		}
+	} else {
+		oldStatus = pr.status
+		pr.Node = n
+		pr.status = StatusAlive
+		pr.leaveTime = time.Time{}
+	}
+
+	p.peers[n.Name] = pr
+
+	if oldStatus == StatusFailed || oldStatus == StatusLeft {
+		level.Debug(p.logger).Log("msg", "peer rejoined", "peer", pr.Node)
+		p.failedPeers = removeOldPeer(p.failedPeers, pr.Name)
+	}
+}
+
+func (p *Peer) peerLeave(n *memberlist.Node) {
+	p.peerLock.RLock()
+	defer p.peerLock.RUnlock()
+
+	pr, ok := p.peers[n.Name]
+	if !ok {
+		// Why are we receiving a leave notification from a node that
+		// never joined?
+		// unknownPeerLeaveCounter
+		return
+	}
+
+	pr.status = StatusFailed
+	pr.leaveTime = time.Now()
+	p.failedPeers = append(p.failedPeers, pr)
+	p.peers[n.Name] = pr
+
+	// TODO: peerLeaveCounter
+	level.Debug(p.logger).Log("msg", "peer left", "peer", pr.Node)
+}
+
+func (p *Peer) peerUpdate(n *memberlist.Node) {
+	p.peerLock.RLock()
+	defer p.peerLock.RUnlock()
+
+	pr, ok := p.peers[n.Name]
+	if !ok {
+		// Why are we receiving an update from a node that never
+		// joined?
+		// unknownPeerUpdateCounter
+		return
+	}
+
+	pr.Node = n
+	p.peers[n.Name] = pr
+
+	// TODO: peerUpdateCounter
+	level.Debug(p.logger).Log("msg", "peer updated", "peer", pr.Node)
 }
 
 func (p *Peer) warnIfAlone(logger log.Logger, d time.Duration) {
@@ -172,6 +330,7 @@ func (p *Peer) AddState(key string, s State) *Channel {
 // Leave the cluster, waiting up to timeout.
 func (p *Peer) Leave(timeout time.Duration) error {
 	close(p.stopc)
+	level.Debug(p.logger).Log("msg", "leaving cluster")
 	return p.mlist.Leave(timeout)
 }
 
@@ -323,184 +482,6 @@ func (c *Channel) Broadcast(b []byte) {
 
 // delegate implements memberlist.Delegate and memberlist.EventDelegate
 // and broadcasts its peer's state in the cluster.
-type delegate struct {
-	*Peer
-
-	logger log.Logger
-	bcast  *memberlist.TransmitLimitedQueue
-
-	messagesReceived     *prometheus.CounterVec
-	messagesReceivedSize *prometheus.CounterVec
-	messagesSent         *prometheus.CounterVec
-	messagesSentSize     *prometheus.CounterVec
-}
-
-func newDelegate(l log.Logger, reg prometheus.Registerer, p *Peer) *delegate {
-	bcast := &memberlist.TransmitLimitedQueue{
-		NumNodes:       p.ClusterSize,
-		RetransmitMult: 3,
-	}
-	messagesReceived := prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "alertmanager_cluster_messages_received_total",
-		Help: "Total number of cluster messsages received.",
-	}, []string{"msg_type"})
-	messagesReceivedSize := prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "alertmanager_cluster_messages_received_size_total",
-		Help: "Total size of cluster messages received.",
-	}, []string{"msg_type"})
-	messagesSent := prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "alertmanager_cluster_messages_sent_total",
-		Help: "Total number of cluster messsages sent.",
-	}, []string{"msg_type"})
-	messagesSentSize := prometheus.NewCounterVec(prometheus.CounterOpts{
-		Name: "alertmanager_cluster_messages_sent_size_total",
-		Help: "Total size of cluster messages sent.",
-	}, []string{"msg_type"})
-	gossipClusterMembers := prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "alertmanager_cluster_members",
-		Help: "Number indicating current number of members in cluster.",
-	}, func() float64 {
-		return float64(p.ClusterSize())
-	})
-	peerPosition := prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "alertmanager_peer_position",
-		Help: "Position the Alertmanager instance believes it's in. The position determines a peer's behavior in the cluster.",
-	}, func() float64 {
-		return float64(p.Position())
-	})
-	healthScore := prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "alertmanager_cluster_health_score",
-		Help: "Health score of the cluster. Lower values are better and zero means 'totally healthy'.",
-	}, func() float64 {
-		return float64(p.mlist.GetHealthScore())
-	})
-	messagesQueued := prometheus.NewGaugeFunc(prometheus.GaugeOpts{
-		Name: "alertmanager_cluster_messages_queued",
-		Help: "Number of cluster messsages which are queued.",
-	}, func() float64 {
-		return float64(bcast.NumQueued())
-	})
-
-	messagesReceived.WithLabelValues("full_state")
-	messagesReceivedSize.WithLabelValues("full_state")
-	messagesReceived.WithLabelValues("update")
-	messagesReceivedSize.WithLabelValues("update")
-	messagesSent.WithLabelValues("full_state")
-	messagesSentSize.WithLabelValues("full_state")
-	messagesSent.WithLabelValues("update")
-	messagesSentSize.WithLabelValues("update")
-
-	reg.MustRegister(messagesReceived, messagesReceivedSize, messagesSent, messagesSentSize,
-		gossipClusterMembers, peerPosition, healthScore, messagesQueued)
-
-	return &delegate{
-		logger:               l,
-		Peer:                 p,
-		bcast:                bcast,
-		messagesReceived:     messagesReceived,
-		messagesReceivedSize: messagesReceivedSize,
-		messagesSent:         messagesSent,
-		messagesSentSize:     messagesSentSize,
-	}
-}
-
-// NodeMeta retrieves meta-data about the current node when broadcasting an alive message.
-func (d *delegate) NodeMeta(limit int) []byte {
-	return []byte{}
-}
-
-// NotifyMsg is the callback invoked when a user-level gossip message is received.
-func (d *delegate) NotifyMsg(b []byte) {
-	d.messagesReceived.WithLabelValues("update").Inc()
-	d.messagesReceivedSize.WithLabelValues("update").Add(float64(len(b)))
-
-	var p clusterpb.Part
-	if err := proto.Unmarshal(b, &p); err != nil {
-		level.Warn(d.logger).Log("msg", "decode broadcast", "err", err)
-		return
-	}
-	s, ok := d.states[p.Key]
-	if !ok {
-		return
-	}
-	if err := s.Merge(p.Data); err != nil {
-		level.Warn(d.logger).Log("msg", "merge broadcast", "err", err, "key", p.Key)
-		return
-	}
-}
-
-// GetBroadcasts is called when user data messages can be broadcasted.
-func (d *delegate) GetBroadcasts(overhead, limit int) [][]byte {
-	msgs := d.bcast.GetBroadcasts(overhead, limit)
-	d.messagesSent.WithLabelValues("update").Add(float64(len(msgs)))
-	for _, m := range msgs {
-		d.messagesSentSize.WithLabelValues("update").Add(float64(len(m)))
-	}
-	return msgs
-}
-
-// LocalState is called when gossip fetches local state.
-func (d *delegate) LocalState(_ bool) []byte {
-	all := &clusterpb.FullState{
-		Parts: make([]clusterpb.Part, 0, len(d.states)),
-	}
-	for key, s := range d.states {
-		b, err := s.MarshalBinary()
-		if err != nil {
-			level.Warn(d.logger).Log("msg", "encode local state", "err", err, "key", key)
-			return nil
-		}
-		all.Parts = append(all.Parts, clusterpb.Part{Key: key, Data: b})
-	}
-	b, err := proto.Marshal(all)
-	if err != nil {
-		level.Warn(d.logger).Log("msg", "encode local state", "err", err)
-		return nil
-	}
-	d.messagesSent.WithLabelValues("full_state").Inc()
-	d.messagesSentSize.WithLabelValues("full_state").Add(float64(len(b)))
-	return b
-}
-
-func (d *delegate) MergeRemoteState(buf []byte, _ bool) {
-	d.messagesReceived.WithLabelValues("full_state").Inc()
-	d.messagesReceivedSize.WithLabelValues("full_state").Add(float64(len(buf)))
-
-	var fs clusterpb.FullState
-	if err := proto.Unmarshal(buf, &fs); err != nil {
-		level.Warn(d.logger).Log("msg", "merge remote state", "err", err)
-		return
-	}
-	d.mtx.RLock()
-	defer d.mtx.RUnlock()
-
-	for _, p := range fs.Parts {
-		s, ok := d.states[p.Key]
-		if !ok {
-			continue
-		}
-		if err := s.Merge(p.Data); err != nil {
-			level.Warn(d.logger).Log("msg", "merge remote state", "err", err, "key", p.Key)
-			return
-		}
-	}
-}
-
-// NotifyJoin is called if a peer joins the cluster.
-func (d *delegate) NotifyJoin(n *memberlist.Node) {
-	level.Debug(d.logger).Log("received", "NotifyJoin", "node", n.Name, "addr", n.Address())
-}
-
-// NotifyLeave is called if a peer leaves the cluster.
-func (d *delegate) NotifyLeave(n *memberlist.Node) {
-	level.Debug(d.logger).Log("received", "NotifyLeave", "node", n.Name, "addr", n.Address())
-}
-
-// NotifyUpdate is called if a cluster peer gets updated.
-func (d *delegate) NotifyUpdate(n *memberlist.Node) {
-	level.Debug(d.logger).Log("received", "NotifyUpdate", "node", n.Name, "addr", n.Address())
-}
-
 func resolvePeers(ctx context.Context, peers []string, myAddress string, res net.Resolver, waitIfEmpty bool) ([]string, error) {
 	var resolvedPeers []string
 
@@ -612,4 +593,15 @@ func retry(interval time.Duration, stopc <-chan struct{}, f func() error) error 
 		case <-tick.C:
 		}
 	}
+}
+
+func removeOldPeer(old []peer, name string) []peer {
+	new := make([]peer, 0, len(old))
+	for _, p := range old {
+		if p.Name != name {
+			new = append(new, p)
+		}
+	}
+
+	return new
 }
