+++ cluster/cluster_test.go
@@ -37,6 +37,7 @@ func TestJoin(t *testing.T) {
 		0*time.Second,
 		0*time.Second,
 		0*time.Second,
+		0*time.Second,
 	)
 	require.NoError(t, err)
 	require.False(t, p == nil)
